my_cd.cをmy_cdディレクトリに移動&一部関数を外に移動
